+++ python/ray/data/datasource/savemode.py
@@ -1,0 +1,8 @@
+from enum import Enum
+
+
+class SaveMode(str, Enum):
+    APPEND = "append"
+    OVERWRITE = "overwrite"
+    IGNORE = "ignore"
+    ERROR = "error"
